Database autoupdater made, bug fixes
Created an auto-updater which will send the batch of read RFID tags to
the database after a period of time has passed. Fixed some null pointer
errors with the DuplicateReadDetector. Added a field to TagWrapper so
that when a tag is wrapped up it saves the location of the RFID reader
that scanned it. This information will be used to properly update the
location of the tag in the database. Edited the TagReporter so it
constructs TagWrappers using the scanned tag and the reader that scanned
the tag. Solved synchronization issues with DBUpdateTimer by having all
functions which alter the tag batch be synchronized (all are located in
DuplicateReadDetector for now).