EPC to UPC converter made
EPC to UPC converter has been created. I'm not 100% sure it works for
all cases. I only tested it with the first excel spreadsheet example and
it worked. I have large portions of code commented out because I've
written a far more concise set of code which will be the final product
after doing some more testing. Basically I'm only keeping the commented
out code as a reference until this is proven to be bug free.