Testing Database
Added some tests for the database. Tests include opening connection,
closing connection, inserting new tags, and updating existing tags.
Added a test class called MyTag.  This class is simply the Impinj Tag
class. There was an issue of the Impinj Tag class having many protected
mutator methods that we need to use, so I made a subclass of that Tag in
the test class with public method versions of the protected classes.