Improving UI and adding second window
Added class for second window. Going to tinker with it for today and see
if I can use it to generate a list to populate DDLs and if I can
populate its Text fields via input list. Need to double check how to pop
this window to the front as well